Move SaveManager's value to Player // Change Item Code
세이브 매니저에서 담당하던 값들을 플레이어로 이전. 세이브 매니저에서 값을 불러오던 함수 등은 PlayerManager로 이전.
Money에서 문제가 생겨 관련 값들 float으로 변경
아이템 코드 규칙 변경에 따른 수정

+++ Assets/Scripts/Item/ProcessedItem.cs
@@ -7,9 +7,9 @@ public class ProcessedItem : Item
     public ItemLevel Level { get; }
     public string Keyword { get; }
     public string FlavorText { get; }
-    public int Price { get; }
+    public float Price { get; }
 
-    public ProcessedItem(int code, string name, ItemLevel level, Sprite spriteImage, string keyword, string flavorText, int price)
+    public ProcessedItem(int code, string name, ItemLevel level, Sprite spriteImage, string keyword, string flavorText, float price)
     {
         Code = code;
         Name = name;
