+++ Assets/Scripts/Item/RawItem.cs
@@ -6,10 +6,10 @@ using System.Linq;
 public class RawItem : Item
 {
     public List<int> OutputCode { get; }
-    public List<int> Price { get; }
+    public List<float> Price { get; }
     public List<float> Duration { get; }
 
-    public RawItem(int code, string name, Sprite spriteImage, List<int> outputCode, List<int> price, List<float> duration )
+    public RawItem(int code, string name, Sprite spriteImage, List<int> outputCode, List<float> price, List<float> duration )
     {
         Code = code;
         Name = name;
