+++ Assets/Scripts/Recipe/Recipe.cs
@@ -5,10 +5,10 @@ public class Recipe
 {
     public int Input { get; private set; }
     public int Output { get; private set;  }
-    public int Price { get; private set; }
+    public float Price { get; private set; }
     public float Duration { get; private set; }
 
-    public Recipe(int input, int output, int price, float duration) {
+    public Recipe(int input, int output, float price, float duration) {
         Input = input;
         Output = output;
         Price = price;
